refactor(netutils): split probe and ftp applets into shared core + per-CLI files
Closes #768
Closes #770

#768 probe: move the shared probe/packet helpers (spec table, validated
Target, injectable transport, target parsing, common Run flow) to probe.go
and give each CLI surface its own file (traceroute.go for traceroute and
traceroute6, ping6.go, arping.go). Behavior unchanged.

#770 ftp: extract the shared FTP transport/session layer (control-connection
dial + greeting, login, binary mode, PASV data channel, RETR/STOR transfers,
PASV parsing, and the shared Run flow) into client.go, leaving ftpget.go and
ftpput.go as thin per-direction CLI surfaces. Behavior unchanged.
