+++ internal/applets/netutils/probe/arping.go
@@ -1,0 +1,5 @@
+package probe
+
+// NewArping returns the arping applet (local-link ARP probe). It shares the spec
+// table, target validation, and privileged transport defined in probe.go.
+func NewArping() *Command { return &Command{kind: kindArping} }
